btrfs-progs: move disk-io.c to kernel-shared/
