+++ btrfs-map-logical.c
@@ -24,7 +24,7 @@
 #include "kerncompat.h"
 #include "ctree.h"
 #include "volumes.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 #include "transaction.h"
 #include "kernel-lib/list.h"
