+++ btrfstune.c
@@ -28,7 +28,7 @@
 
 #include "kerncompat.h"
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "transaction.h"
 #include "common/utils.h"
 #include "volumes.h"
