+++ cmds/filesystem-usage.c
@@ -30,7 +30,7 @@
 #include "common/string-table.h"
 #include "cmds/filesystem-usage.h"
 #include "cmds/commands.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 
 #include "version.h"
 #include "common/help.h"
