+++ cmds/inspect-dump-super.c
@@ -23,7 +23,7 @@
 #include <getopt.h>
 
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 #include "common/utils.h"
 #include "cmds/commands.h"
