+++ cmds/inspect-tree-stats.c
@@ -28,7 +28,7 @@
 
 #include "kerncompat.h"
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 #include "transaction.h"
 #include "kernel-lib/list.h"
