+++ cmds/rescue-chunk-recover.c
@@ -33,7 +33,7 @@
 #include "kernel-lib/radix-tree.h"
 #include "ctree.h"
 #include "common/extent-cache.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "volumes.h"
 #include "transaction.h"
 #include "crypto/crc32c.h"
