+++ cmds/rescue-super-recover.c
@@ -28,7 +28,7 @@
 
 #include "kerncompat.h"
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "kernel-lib/list.h"
 #include "common/utils.h"
 #include "crypto/crc32c.h"
