+++ cmds/rescue.c
@@ -22,7 +22,7 @@
 #include "ctree.h"
 #include "volumes.h"
 #include "transaction.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "cmds/commands.h"
 #include "common/utils.h"
 #include "common/help.h"
