+++ cmds/restore.c
@@ -38,7 +38,7 @@
 #include <sys/xattr.h>
 
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 #include "transaction.h"
 #include "kernel-lib/list.h"
