+++ cmds/scrub.c
@@ -41,7 +41,7 @@
 #include "ioctl.h"
 #include "common/utils.h"
 #include "volumes.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 
 #include "cmds/commands.h"
 #include "common/help.h"
