+++ common/utils.c
@@ -45,7 +45,7 @@
 #include "kerncompat.h"
 #include "kernel-lib/radix-tree.h"
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "transaction.h"
 #include "crypto/crc32c.h"
 #include "common/utils.h"
