+++ convert/common.c
@@ -16,7 +16,7 @@
 
 #include <unistd.h>
 #include <uuid/uuid.h>
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "volumes.h"
 #include "common/utils.h"
 #include "common/path-utils.h"
