+++ convert/main.c
@@ -92,7 +92,7 @@
 #include <stdbool.h>
 
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "volumes.h"
 #include "transaction.h"
 #include "common/utils.h"
