+++ file.c
@@ -19,7 +19,7 @@
 #include <sys/stat.h>
 #include "ctree.h"
 #include "common/utils.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "transaction.h"
 #include "kerncompat.h"
