+++ kernel-lib/raid56.c
@@ -25,7 +25,7 @@
 #include <unistd.h>
 #include "kerncompat.h"
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "volumes.h"
 #include "common/utils.h"
 #include "kernel-lib/raid56.h"
