+++ kernel-shared/disk-io.c
@@ -26,7 +26,7 @@
 #include "kerncompat.h"
 #include "kernel-lib/radix-tree.h"
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "volumes.h"
 #include "transaction.h"
 #include "crypto/crc32c.h"
