+++ kernel-shared/disk-io.h
