+++ kernel-shared/file-item.c
@@ -21,7 +21,7 @@
 #include "kerncompat.h"
 #include "kernel-lib/radix-tree.h"
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "transaction.h"
 #include "kernel-shared/print-tree.h"
 #include "crypto/crc32c.h"
