+++ kernel-shared/free-space-cache.c
@@ -20,7 +20,7 @@
 #include "ctree.h"
 #include "kernel-shared/free-space-cache.h"
 #include "transaction.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "kernel-shared/extent_io.h"
 #include "crypto/crc32c.h"
 #include "kernel-lib/bitops.h"
