+++ kernel-shared/inode-item.c
@@ -17,7 +17,7 @@
  */
 
 #include "ctree.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "transaction.h"
 
 static int find_name_in_backref(struct btrfs_path *path, const char * name,
