+++ kernel-shared/root-tree.c
@@ -18,7 +18,7 @@
 
 #include "ctree.h"
 #include "transaction.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 
 int btrfs_find_last_root(struct btrfs_root *root, u64 objectid,
