+++ transaction.c
@@ -15,7 +15,7 @@
  */
 
 #include "kerncompat.h"
-#include "disk-io.h"
+#include "kernel-shared/disk-io.h"
 #include "transaction.h"
 #include "kernel-shared/delayed-ref.h"
 #include "common/messages.h"
